Downloader.py added and bugs fixed
